\name{ACCQR}
\alias{ACCQR}
\alias{dmatrix_ACCQR}
\alias{genmatrix}
\alias{CCompQreg}
\alias{AdpCComptQreg}
\alias{rho_tao_ACCQR}
\alias{HBIC_ACCQ}
\alias{findlambda_ACCQ}
\title{ACCQR}
\description{
compositional composite quantile regression with adptive lasso, using HBIC to select tuning parameter
}
